requirements.txt
langchain==0.0.228
chromadb==0.3.26
PyMuPDF==1.22.5
python-dotenv==1.0.0
unstructured==0.8.0
tqdm==4.65.0
flask==2.3.2
flask-cors==4.0.0
slack-sdk==3.21.3
openai==0.27.8
tiktoken==0.4.0
pymongo==4.4.1
